requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
sentence-transformers
google-generativeai
sse-starlette
numpy
